+++ src/Dreamberd/Bytecode/Encode.hs
@@ -1,6 +1,7 @@
 module Dreamberd.Bytecode.Encode (
     transpileCall,
     transpileInt,
+    transpileInteger,
     transpileString,
     transpileValue,
     transpileInstruction,
@@ -12,10 +13,21 @@ import Data.Char (chr)
 import Dreamberd.Vm (Call (..), EnvValue (..), Insts (..), Value (..))
 import System.Endian (Endianness (..), getSystemEndianness)
 
+getIntegerSize :: Integer -> Int
+getIntegerSize nb = if nb == 0 then 0 else 1 + getIntegerSize (nb `shiftR` 8)
+
 transpileInt :: Int -> [Char]
 transpileInt nb = case getSystemEndianness of
-    LittleEndian -> map chr $ reverse $ map (`mod` 256) $ take 4 $ iterate (`shiftR` 8) nb
-    BigEndian -> map (chr . (`mod` 256)) (take 4 $ iterate (`shiftR` 8) nb)
+    LittleEndian -> map chr $ reverse $ map (`mod` 256) $ take 8 $ iterate (`shiftR` 8) nb
+    BigEndian -> map (chr . (`mod` 256)) (take 8 $ iterate (`shiftR` 8) nb)
+
+transpileInteger :: Integer -> [Char]
+transpileInteger nb =
+    (if nb >= 0 then toEnum 0x51 else toEnum 0x52) : transpileInt size ++ case getSystemEndianness of
+        LittleEndian -> map chr $ reverse $ map (`mod` 256) $ take size $ iterate (`shiftR` 8) (fromInteger (abs nb))
+        BigEndian -> map (chr . (`mod` 256)) (take size $ iterate (`shiftR` 8) (fromInteger (abs nb)))
+  where
+    size = getIntegerSize (abs nb)
 
 transpileString :: String -> [Char]
 transpileString str = transpileInt (length str) ++ str
@@ -26,10 +38,13 @@ transpileCall (FunctionName str) = toEnum 0x22 : transpileString str
 
 transpileValue :: Value -> [Char]
 transpileValue (Number nb) = toEnum 0x11 : transpileInt nb
-transpileValue (Bool b) = toEnum 0x12 : (if b then [toEnum 1] else [toEnum 0])
-transpileValue (String str) = toEnum 0x13 : transpileString str
-transpileValue (Symbol call) = toEnum 0x14 : transpileCall call
-transpileValue Void = [toEnum 0x15]
+transpileValue (Float nb) = toEnum 0x12 : transpileInteger l ++ transpileInt r
+  where
+    (l, r) = decodeFloat nb
+transpileValue (Bool b) = toEnum 0x13 : (if b then [toEnum 1] else [toEnum 0])
+transpileValue (String str) = toEnum 0x14 : transpileString str
+transpileValue (Symbol call) = toEnum 0x15 : transpileCall call
+transpileValue Void = [toEnum 0x16]
 
 transpileEnvValue :: Maybe EnvValue -> [Char]
 transpileEnvValue Nothing = [toEnum 0x53]