+++ src/Lib.hs
@@ -5,6 +5,7 @@ import Dreamberd.Lib (compileDreamberdCode, compileToAst, compileToVm, executeDr
 import Input (processFile, processInput)
 import Lisp.Lib (evaluateAndPrintLispResult)
 
+import Dreamberd.Types (File (..))
 import Options.Applicative
 
 glados :: IO ()
@@ -17,6 +18,6 @@ runGlados (Compile file _ False True) = processFile compileToVm file
 runGlados (Compile file output False False) = processFile (`compileDreamberdCode` output) file
 runGlados (Execute file) = processFile executeDreamberdBytecode file
 runGlados (Run file) = processFile runDreamberdCode file
-runGlados (Lisp (Just file)) = processFile evaluateAndPrintLispResult file
+runGlados (Lisp (Just file)) = processFile (\(File _ sourceCode) -> evaluateAndPrintLispResult sourceCode) file
 runGlados (Lisp Nothing) = processInput evaluateAndPrintLispResult
 runGlados Version = putStrLn "GLaDOS v1.0.0"