+++ test/functional/run.py
@@ -7,16 +7,14 @@ from difflib import unified_diff as diff
 from glob import glob
 import re
 
-LANGUAGES = [("Lisp", ".lsp", "lisp", "lisp"), ("Dreamberd", ".db4", "dreamberd", "-c")]
+LANGUAGES = [("Lisp", ".lsp", "lisp", "lisp", ""), ("Dreamberd Compile + Execute", ".db4", "dreamberd", "compile", ";./glados execute a.out"), ("Dreamberd run", ".db4", "dreamberd", "run", "")]
 
 def get_test_paths(folder_name: str, extension: str) -> list[str]:
     folder = path.dirname(path.realpath(__file__))
     return glob(folder + f'/src/{folder_name}/*{extension}', recursive=True)
 
-
-def run_command(test_path: str, binary: str) -> CompletedProcess[str]:
-    return run(f'{binary} {test_path}', shell=True, capture_output=True, text=True)
-
+def run_command(test_path: str, binary: str, after_command: str) -> CompletedProcess[str]:
+    return run(f'{binary} {test_path} {after_command}', shell=True, capture_output=True, text=True)
 
 def disp_err(output: CompletedProcess[str], expected_output: str, expected_error: str) -> None:
     if output.stdout != expected_output:
@@ -32,9 +30,8 @@ def disp_err(output: CompletedProcess[str], expected_output: str, expected_error
             print(line, end="")
         print()
 
-
-def run_test(test_path: str, extension: str, flag: str, is_debug: bool, is_full_log: bool, has_color: bool) -> bool:
-    output = run_command(test_path, f"./glados {flag}")
+def run_test(test_path: str, extension: str, flag: str, after_command: str, is_debug: bool, is_full_log: bool, has_color: bool) -> bool:
+    output = run_command(test_path, f"./glados {flag}", after_command)
     expected_output, expected_error = '', ''
 
     if path.exists(test_path.replace(extension, ".out")):
@@ -53,6 +50,7 @@ def run_test(test_path: str, extension: str, flag: str, is_debug: bool, is_full_
         err = f'Test "{name}" failed.'
         print(colored(err, "red") if has_color else err)
         return False
+
     if is_full_log:
         success = f'Test "{name}" passed.'
         print(colored(success, "green") if has_color else success)
@@ -66,18 +64,46 @@ if __name__ == "__main__":
 
     total_tests, total_passed = 0, 0
 
-    for name, extension, folder, flag in LANGUAGES:
-        print(f"Running {name} tests:")
-        nb_passed, nb_failed = 0, 0
-        tests = get_test_paths(folder_name=folder, extension=extension)
-        nb_tests = len(tests)
+    for name, extension, folder, flag, after_command in LANGUAGES:
+         print(f"\n\n===================================================================================================================" + (len(name) - 6) * "=")
+         print(f"================================================== TEST [{name}] ==================================================")
+         print(f"===================================================================================================================" + (len(name) - 6) * "=" + "\n")
+         nb_passed, nb_failed = 0, 0
+         test_passed, test_failed = [], []
+         tests = get_test_paths(folder_name=folder, extension=extension)
+         nb_tests = len(tests)
 
-        for test in tests:
-            passed = run_test(test, extension, flag, is_debug, is_full_log, has_color)
-            nb_passed += 1 if passed else 0
-            nb_failed += 1 if not passed else 0
-        print(f"Ran {nb_tests} tests ({nb_passed} passed and {nb_failed} failed).\n")
-        total_tests += nb_tests
-        total_passed += nb_passed
+         for test in tests:
+             name = re.findall('[^/]*$', test)[0]
+             passed = run_test(test, extension, flag, after_command, is_debug, is_full_log, has_color)
+             nb_passed += 1 if passed else 0
+             test_passed.append(name) if passed else test_failed.append(name)
+             nb_failed += 1 if not passed else 0
+         print(f"\n##################################################################################################################")
+         print(f"#================================================= RESULTS ======================================================#")
+         print(f"#================================================================================================================#")
+         print(f"#                                              Ran {nb_tests} tests")
+         print(f"# {nb_passed} passed :")
+         for test in test_passed:
+            print(f"#   - {colored(test, 'green')}")
+         print(f"#\n# {nb_failed} failed :")
+         for test in test_failed:
+            print(f"#   - {colored(test, 'red')}")
+         print(f"#")
+         success_percentage = (nb_passed / nb_tests) * 100
+         failure_percentage = (nb_failed / nb_tests) * 100
+
+         success_percentage = round(success_percentage, 2)
+         failure_percentage = round(failure_percentage, 2)
+
+         success_bar = int(success_percentage) * '#'
+         failure_bar = int(failure_percentage) * '#'
+
+         print(f"# Success rate : {colored(success_percentage, 'green')}%")
+         print("# "+colored(success_bar, "green") + colored(failure_bar, "red"))
+         print(f"#")
+         print(f"#\n##################################################################################################################\n")
+         total_tests += nb_tests
+         total_passed += nb_passed
 
     exit(nb_tests - nb_passed)